stack/provider: add PrimitiveSlice for wrapping slice of primitives
If Slice is constructed from e.g. a slice of ints []int, then
hcl.DecodeObject fails with:

  root.marathon_app.app.ports: not an object type for map (*ast.ListType)

This PR creates another workaround for primitives - PrimitiveSlice.
The new type wraps each primitive p with map[string]interface{}{"": p},
which makes hcl consider the slice as *ast.ObjectType instead
if *ast.ListType.
